requirements.txt
google-generativeai
pandas
kaggle
python-dotenv
ipykernel
nbformat
tqdm
bs4